Put all errors in their own place
